fix: change error handling path
The gist of this change is to move the fallible logic of attaching ebpf to an device out of the spawned task watcher. This allows to not add the task to the task list if it fails before it. Also it prevents the hot loop from just early exiting on encountered errors.